ux improvements again
