+++ src/styles/main.css
@@ -1,6 +1,4 @@
-/* ---------------------------------------------------------
-   1) GLOBAL BACKGROUND (PROPER OPACITY)
-----------------------------------------------------------*/
+/* 1) GLOBAL RESET & FONT */
 html, body {
   margin: 0;
   padding: 0;
@@ -8,34 +6,31 @@ html, body {
   width: 100%;
   height: 100%;
   font-family: sans-serif;
+}
 
+/* 2) GLOBAL BACKGROUND */
+body {
   /* Semi-transparent warm overlay plus background image */
-  background: 
-    linear-gradient(rgba(255, 248, 240, 0.6), rgba(255, 248, 240, 0.6)), 
+  background:
+    linear-gradient(rgba(255, 248, 240, 0.6), rgba(255, 248, 240, 0.6)),
     url("https://i2.pickpik.com/photos/614/325/207/tuscany-villa-italian-nature-preview.jpg")
     no-repeat center center fixed;
   background-size: cover;
-
-  /* Prevent scrolling */
-  overflow: hidden;
 }
 
-/* ---------------------------------------------------------
-   2) GLOBAL VARIABLES (USABLE IN COMPONENTS)
-----------------------------------------------------------*/
+/* 3) GLOBAL VARIABLES */
 :root {
-  --header-height: 60px;           /* Default header height */
-  --footer-height: 20px;           /* Default footer height */
-  --date-nav-buttons-height: 75px; /* Height for oldest/select/newest nav */
-  
-  /* Header & Global Layout */
-  --header-bg: #A67B5B;
-  --header-text-color: #ecf0f1;
+  --header-height: 60px;
+  --footer-height: 20px;
   --global-padding: 1rem;
 
+  /* Warm terracotta theme for header/nav */
+  --header-bg: #A67B5B; /* or your chosen terracotta color */
+  --header-text-color: #fff; /* Contrasts with the warm background */
+
   /* Navigation & button colors */
-  --nav-link-color: #ecf0f1;
-  --nav-link-hover-color: #e74c3c;
+  --nav-link-color: #fff;
+  --nav-link-hover-color: #b22222; /* slightly lighter to contrast */
 
   /* Help button gradients */
   --help-gradient-default: linear-gradient(135deg, #ff9a9e, #ffe0cc);
@@ -45,4 +40,9 @@ html, body {
   /* Login button colors */
   --login-border-color: #fff;
   --login-hover-bg: rgba(255, 255, 255, 0.2);
+}
+
+/* 4) OPTIONAL: Hide overflow if you don’t want scrolling */
+html, body {
+  overflow: hidden;
 }
